日光の強度を移動
