+++ src/components/atoms/SunStrengthSelect/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './index.scss';
-import { 日光の強度 } from '@frontend/domain/model/鉢';
 import { Select, SelectProps } from 'antd';
+import { 日光の強度 } from 'src/domain/model/鉢/日光の強度';
 
 export namespace 日光の強度Select {
   export type Props = {
@@ -14,8 +14,8 @@ export namespace 日光の強度Select {
 }
 
 const 指定なし = '';
-const 日光の強度一覧 = [...Object.values(日光の強度), 指定なし];
-const options = 日光の強度一覧.map(i => ({ value: i, label: i }));
+const 日光の強度一覧 = [...Object.values(日光の強度), { name: 指定なし }];
+const options = 日光の強度一覧.map(i => ({ value: i.name, label: i.name }));
 
 export const 日光の強度Select: React.FC<日光の強度Select.Props> = props => {
   const { onChange, value, isLoading, size, placeholder } = props;
