+++ src/components/molecules/FilterInput/index.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import './index.scss';
 import { useFilter } from '@frontend/store/filter/action';
 import { Col, Input, InputNumber, InputNumberProps, InputProps, Row, Slider } from 'antd';
-import { 日光の強度 } from '@frontend/domain/model/鉢';
 import { SliderRangeProps } from 'antd/es/slider';
 import { optionalValue } from '@frontend/supports/functions';
 import { 日光の強度Select } from '@frontend/components/atoms/SunStrengthSelect';
@@ -10,6 +9,7 @@ import { SYMBOL_ICONS } from '@frontend/supports/icons';
 import cn from 'classnames';
 import { DebugOnly } from '@frontend/components/atoms/DebugOnly';
 import { onClickOrEnter, onKeyEnter } from '@frontend/supports/keyboardAction';
+import { 日光の強度 } from 'src/domain/model/鉢/日光の強度';
 
 namespace フィルタ条件の入力 {
   export type Props = {};
