+++ src/components/organisms/TableView/index.tsx
@@ -118,6 +118,7 @@ const Row: React.FC<{ 棚: 棚; user: User | undefined }> = props => {
     <>
       <Table
         key={棚.id}
+        // @ts-ignore
         columns={columns}
         dataSource={鉢一覧.map(鉢 => {
           return Object.assign({ ...鉢 }, { key: 鉢.id });
@@ -152,6 +153,7 @@ export const テーブル表示: React.FC<TableViewProps> = props => {
         columns={columns}
         expandable={{ expandedRowRender: expandedRowRender(user) }}
         pagination={undefined}
+        // @ts-ignore
         dataSource={棚一覧.map(棚 => {
           return Object.assign({ ...棚 }, { key: 棚.id });
         })}
