+++ src/components/organisms/植物ごとのデフォルト設定モーダル/新規作成/index.tsx
@@ -3,7 +3,6 @@ import './index.scss';
 import { useWithLoading } from '@frontend/supports/ui';
 import { useAuthState } from '@frontend/store/auth/action';
 import { Modal, ModalProps } from 'antd';
-import { 日光の強度, 日光の強度設定 } from '@frontend/domain/model/鉢';
 import { Control, useController, UseControllerReturn, useForm } from 'react-hook-form';
 import { MyInputWithAlert } from '@frontend/components/atoms/MyInputWithAlert';
 import { MyFormLayout } from '@frontend/components/molecules/MyForm';
@@ -12,6 +11,8 @@ import { モーダルの見出し } from '@frontend/components/atoms/ModalTitle'
 import { 季節 } from '@frontend/domain/const/季節';
 import { 植物ごとのデフォルト設定 } from '@frontend/domain/model/植物のデフォルト設定';
 import { 日光の強度Select } from '@frontend/components/atoms/SunStrengthSelect';
+import { 日光の強度設定 } from 'src/domain/model/鉢/日光の強度設定';
+import { 日光の強度 } from 'src/domain/model/鉢/日光の強度';
 
 export namespace 植物ごとのデフォルト設定の新規作成モーダル {
   export type Ref = {
