+++ src/components/organisms/植物ごとのデフォルト設定モーダル/編集/index.tsx
@@ -1,13 +1,13 @@
 import React, { forwardRef, useImperativeHandle, useState } from 'react';
 import './index.scss';
 import { Descriptions, Modal, ModalProps } from 'antd';
-import { 日光の強度 } from '@frontend/domain/model/鉢';
 import { 植物ごとのデフォルト設定Id } from '@frontend/domain/model/植物のデフォルト設定';
 import { 季節 } from '@frontend/domain/const/季節';
 import { 日光の強度Select } from '@frontend/components/atoms/SunStrengthSelect';
 import { 育成タイプSelect } from '@frontend/components/atoms/GrowthTypeSelect';
 import { MyEditable } from '@frontend/components/atoms/Editable';
 import { use植物ごとのデフォルト設定 } from '@frontend/store/master/action';
+import { 日光の強度 } from 'src/domain/model/鉢/日光の強度';
 
 export namespace 植物ごとのデフォルト設定編集モーダル {
   export type Ref = {
