+++ src/components/pages/植物ごとのデフォルト設定ページ/index.tsx
@@ -8,9 +8,9 @@ import { MyEditable } from '@frontend/components/atoms/Editable';
 import { 植物ごとのデフォルト設定 } from '@frontend/domain/model/植物のデフォルト設定';
 import { 季節 } from '@frontend/domain/const/季節';
 import { 日光の強度Select } from '@frontend/components/atoms/SunStrengthSelect';
-import { 日光の強度 } from '@frontend/domain/model/鉢';
 import { uniqueArray } from '@frontend/supports/array';
 import { 育成タイプSelect } from '@frontend/components/atoms/GrowthTypeSelect';
+import { 日光の強度 } from 'src/domain/model/鉢/日光の強度';
 
 export namespace 植物ごとのデフォルト設定ページ {
   export type Props = {};
